requirements.txt
discord.py~=2.1.0
table2ascii